Comment out more unused scripts trying to bugfix
My deployed page (jal90.github.io) renders a <span> that does not
exist on my local page. Deleting it in the DOM does nothing - it re-
appears a moment later! It's a thin white bar in the upper left hand
corner of the page.

Since it doesn't exist on my local site, only on the deployed site, I
have to make a bunch of commits and then deploy it so I can see if it's
fixed.